Summer_2024/Communication/Kafka_Demo/consumer/requirements.txt
kafka-python
python-dotenv==1.0.0
pymongo==4.4.1